+++ src/main/java/com/minsoo/co/tireerpserver/model/dto/tire/dot/TireDotSimpleResponse.java
@@ -18,7 +18,7 @@ public class TireDotSimpleResponse {
     private String dot;
 
     @Schema(name = "retail_price", description = "소비자금액")
-    private Long retailPrice;
+    private Integer retailPrice;
 
     private TireDotSimpleResponse(TireDot tireDot) {
         this.tireDotId = tireDot.getId();
